+++ src/main/java/step3/car/game/CarGame.java
@@ -1,14 +1,11 @@
-package step3.game;
-
-import step3.entity.Cars;
-import step3.view.ResultView;
+package step3.car.game;
 
 import java.util.stream.IntStream;
 
 public class CarGame {
     private final Cars cars;
     private final int moves;
-    private final ResultView view;
+    private final CarsView view;
 
     public CarGame(int numberOfCars, int numberOfMoves) {
         if (numberOfCars <= 0 || numberOfMoves <= 0) {
@@ -16,7 +13,7 @@ public class CarGame {
         }
         cars = new Cars(numberOfCars);
         moves = numberOfMoves;
-        view = new ResultView();
+        view = new CarsView();
     }
 
     public void start() {
