+++ src/test/java/step3/integration/CarGameTest.java
@@ -1,7 +1,7 @@
 package step3.integration;
 
 import org.junit.jupiter.api.Test;
-import step3.game.CarGame;
+import step3.car.game.CarGame;
 
 public class CarGameTest {
